Separate json encoder from __init__.py
